+++ lib/status_protocol.dart
@@ -1,6 +1,8 @@
 export 'src/models/status_config.dart';
+export 'src/models/contact_info.dart';
 export 'src/data/status_data.dart';
 export 'src/utils/status_resolver.dart';
 export 'src/widgets/status_screen.dart';
 export 'src/widgets/status_banner.dart';
 export 'src/widgets/status_theme.dart';
+export 'src/widgets/contact_section.dart';